registro

+++ dom.js
@@ -3,18 +3,59 @@ function registrarse (){
     window.location.assign('registro.html')
 }
 
+let password = document.getElementById('contraseña');
+let confirmarDiv = document.getElementById('confirm-container');
+
+password.addEventListener('input', function() {
+    if (password.value.length > 0) {
+        confirmarDiv.style.display = 'block';
+    } else {
+        confirmarDiv.style.display = 'none';
+    }
+});
+
 function registro (){
     let user = document.getElementById('usuario').value
     let nombre = document.getElementById('nombre').value
     let apellido = document.getElementById('apellido').value
     let code = document.getElementById('codigoFamiliar').value
-    let password = document.getElementById('contraseña').value
+    let confirmación = document.getElementById('confirmar');
 
-    usuarioGuardado={user,nombre,apellido,password,code}
+    let num = false
+    let esp = false
+    let nums = '0123456789'
+    let esps = '/?.><,;:][}{=_-+*&^%$#@!~'
+
+    for (let i = 0; i < password.value.length; i++){
+        for (let j = 0; j < nums.length; j++){
+            console.log(password.value[i], nums[j])
+            if (password.value[i] == nums[j]){
+                num = true
+                console.log(num)
+            } 
+        }
+        for (let k = 0; k < nums.length; k++){
+            console.log(password.value[i], esps[k])
+            if (password.value[i] == esps[k]){
+                esp = true
+                console.log(esp)
+            } 
+        }
+    }
+
+    if (code.length != 6){
+        alert('Ingrese codigo de 6 digitos')
+    } else if (num == false || esp == false){
+        alert('Incluya al menos un número y un caracter especial.')
+    } else if (num == true && esp == true && password.value == confirmación.value && code.length == 6){
+        alert(`Bienvenido, ${user}`)
+        window.location.assign('inicio.html')
+        localStorage.setItem('user', user)
+    } else {
+        alert('Contraseñas no coincidentes.')
+    }
     
-    alert(`Bienvenido, ${user}`)
-    window.location.assign('inicio.html')
-    localStorage.setItem('user', user)
+    
 }
 
 function loguearse (){
@@ -90,7 +131,7 @@ function validarModificaciones (){
     let apNew = parseFloat(document.getElementById('apellidoNuevo').value);
     let passNew = document.getElementById('contraseñaNueva').value.trim();
 
-    if (nombreNew.length < 5 || apNew.length < 1 || passNew.length < 1) {
+    if (nombreNew.length < 1 || apNew.length < 1 || passNew.length < 1) {
         alert('Complete los datos!');
     } else {
         modificarUser(nombreNew, apNew, passNew);
